refactor: util 패키지 추가
